Add join/resume game feature for player
Added join_game view which redirects user to game with code given in
GameAccessView.

The view searches for game containing password provided by
GameAccessView form. If such game exists, the view searches for existing
players to check if player had already joined before. If not, the player
is created and added to game with default stats, only if the player
number doesn't exceed the given global maximum number of players per
game, which is defined in game.models.MAX_PLAYERS.